pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ghostkitty-bitcrusher"
version = "2.0.0"
description = "Audio bitcrusher with real-time preview"
authors = [{name = "CATHOUSEMP3"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "numpy>=1.21.0",
    "scipy>=1.7.0",
    "pygame-ce>=2.4.0",
    "customtkinter>=5.0.0",
    "Pillow>=9.0.0",
    "soundfile>=0.10.0",
]

[project.urls]
Homepage = "https://github.com/ghostkitty/bitcrusher"
Repository = "https://github.com/ghostkitty/bitcrusher.git"
Issues = "https://github.com/ghostkitty/bitcrusher/issues"

[project.scripts]
ghostkitty-bitcrusher = "ghostkitty_bitcrusher.main:main"
